Update .Rbuildignore so that only files in the root of the repository are ignored. Otherwise the template files in inst/infrastructure are also removed when the package is built with `R CMD build`.
For an explanation on which files are removed, see the Writing R
Extensions manual:

https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Package-structure

To solve the isssue, I copied the strategy from devtools, which also has
template files in inst/.

https://github.com/hadley/devtools/blob/1ce84b04568ff7846c3da754f28e7e22a23c8737/.Rbuildignore

The .Rbuildingore file uses Perl-like regular expressions. I find
?regexp completely incomprehensible. The link below was much more
straightforward.

. -> match any one character
.* -> match any character zero or more times
\. -> An actual period
^ -> beginning of string
$ -> end of string

https://www.cs.tut.fi/~jkorpela/perl/regexp.html

The .nojekyll file survives the build process, but not .gitignore. It'd
probably just be easier to create it from scratch.