Add NEWS file and fix warnings from R CMD CHECK (ran with `devtools::check()`).
The most confusing thing was rmarkdown. It is in Depends so that it is
loaded and attached when workflowr is. However, since workflowr could be
loaded but not attached, then rmarkdown would not be available. The
manual way to do this would be to add rmarkdown to the NAMESPACE, but
this is automatically generated. Thus I added it to the roxygen docs for
each function that uses any rmarkdown command. I still use rmarkdown::
for clarity, but it is not actually necessary to do so anymore.