Add FAQ about the differences between `wflow_build()` and RStudio Knit HTML.
This was inspired by an example Matthew found. A function in the package
he was developing was relying on a variable defined in the Rmd file
being knit (by accident). He was able to build the HTML with the Knit
HTML button, but did not notice his mistake until running
`wflow_build()`, which did not let the package function use the variable
defined in the global environment.