Explicitly call workflowr:::normalizePath in unit tests.
testthat already uses the normalizePath in the workflowr namespace
during the tests. The reason for explicitly prepending the namespace is
for 1) clarity, so that readers know the base version of the function is
not being used, and 2) so that the unit tests can be copied directly
into the R console for interactive testing.