Specify the global environment when building R Markdown file in an external R process.
Back story: The goal of build_rmd_external is to run render_site such
that it is completely isolated from the current R environment, similar
to how the RStudio "Knit" button works. This is an important feature for
reproducibility.

Matthew accidentally found an edge case where build_rmd_external and
RStudio Knit diverge. He had written a package function that referenced
a variable it hadn't defined inside the function, but it was defined
earlier in the R Markdown document that called the function. Using
RStudio Knit, the package function was able to access the variable
defined earlier in the R Markdown document. However, using
build_rmd_external via wflow_build, it failed due to the undefined
variable. This was good in Matthew's case, since it caught his mistake.
However, I want workflowr to behave identically to RStudio Knit or
manually running render_site in a new R session.

I wasn't sure how to fix this, but then I found this Issue on the
rmarkdown GitHub repo about using callr. The trick is to set the
environment to .GlobalEnv. This is slightly confusing, but it's not
referring to the global environment of the user, but the global
environment of the external R process that is started.

https://github.com/rstudio/rmarkdown/issues/1204