Fix wflow_git_config() on Windows.
Related to Issue #76.

Git writes the global configuration settings to ~/.gitconfig. The
problem is that on Windows, ~ is mapped to ~/Documents. git2r is able
to use the settings in ~/Documents/.gitconfig, but Git will break at the
command line if ~/.gitconfig does not exist. If both exist, then git2r
uses the settings in ~/Documents/.gitconfig. The workaround is to create
an empty ~/.gitconfig first. git2r then writes to this file instead of
creating ~/Documents/.gitconfig.

https://cran.r-project.org/bin/windows/base/rw-FAQ.html#What-are-HOME-and-working-directories_003f