Fix some reasonable errors in units tests.
The wflow_open errors were caused by trying to use relative paths when
the working directory was changing. Surprising this wasn't caught on
Linux, but was caught on both macOS and Windows.

The wflow_convert error was due to the symlink not being resolved since
the file didn't yet exist, which would only manifest on macOS which uses
symlinks in the temp directory.