Simplify get_home.
get_home was failing on AppVeyor b/c HOMEPATH was set to \\. I was
expecting it to be either "\\Users\\appveyor" or "". The nice thing is
that this exposed a bug in which stripping the trailing slash from the
drive resulted in the current working directory being returned.

Here's the hack to skip Travis and not AppVeyor. Put [ci skip] in the
body. AppVeyor only skips if it is in the first line.