Do not treat warnings as errors for devel version of R.
I am getting a warning when building with the development version of R.
The problem file is one created by R itself during the build! Presumably
this will eventually get worked out, but in the meantime I will be more
lenient only for the build with r-devel.

checking serialized R objects in the sources ... WARNING
Found file(s) with version 3 serialization:
  ‘build/vignette.rds’
Such files are only readable in R >= 3.5.0.
Recreate them with R < 3.5.0 or save(version = 2) or saveRDS(version =

https://travis-ci.org/jdblischak/workflowr/jobs/326496597#L1067

I suspect it has to do with this recent commit:

https://github.com/wch/r-source/commit/d9530001046a582ff6a43ca834d6c3586abd0a97