Make paths relative to Git repo when adding files in wflow_commit.
If a file or directory no longer exists, git2r can't accept a relative
path to it. It requires an absolute path or a path relative to the
root of the Git repo.

This was discovered by Peter after the introduction of the feature of
removing unused figures. He had a file that had a figure but then had
zero. Since the directory no longer existed, if wflow_publish was run
from any directory other than the root, then the figure file was not
included in the commit. wflow_remove was designed to handle this since
it is always deleting non-existent files, but wflow_commit also needs
to be able to handle this.

https://github.com/jdblischak/workflowr/blob/a2fa789241d7245170043c07e5af00bd8ccd6b3b/R/wflow_remove.R#L169