Ensure valid final line in file for wflow_remove test.
Strangely a wflow_publish call setting up the unit tests for
wflow_remove failed only on r-hub Ubuntu. The only thing I
could find was that it did not have a final end line b/c of
how I used `cat` (this creates a warning when reading with readLines),
which is how I discovered this). Hopefully this fixes the issue.