Update _site.yml and .nojekyll for use with wflow_site. Close #72.
rmarkdown:::default_site copies all files from the input directory
to the output directory. I never liked that behavior, and now
wflow_site does not do this. It only copies the HTML, figures, and
supporting files for the given R Markdown filing being built.

https://github.com/rstudio/rmarkdown/blob/91d071ffc04025fa03855fa61c208a40b0011845/R/render_site.R#L321

This means that the fields include/exclude are not used, so I
removed them.

https://rmarkdown.rstudio.com/rmarkdown_websites.html#included-files

Also, now that .nojekyll is no longer copied, it needs to be
created by wflow_build. It also needs to be committed by
wflow_publish, which fixes the old bug in #72 where the .nojekyll
file was not committed after changing the output directory.