Replace file.remove() with fs::file_delete().
I decided not to replace unlink() with dir_delete(). This is because
unlink() does not fail if the directory does not exist. My code uses
this all the time. It deletes a directory, e.g. the figures/ directory,
without first checking if it exists. Adding in this extra logic would
take some effort, and at this point I don't think it is worth the
effort.