Attempt to install libgit2 with Homebrew. [skip appveyor]
git2r 0.24.0 was released a few days ago (2019-01-07). There was no big
change in its installation process, however, CRAN no longer builds
binaries for R 3.4.4 (currently set as oldrel on Travis). Thus it tries
to compile libgit2 and fails.

https://cran.r-project.org/package=git2r/NEWS
https://travis-ci.org/jdblischak/workflowr/jobs/477017411#L152
https://travis-ci.org/jdblischak/workflowr/jobs/477017411#L228

My goal is to install libgit2 via Homebrew only for the oldrel build and
not the release (since workflowr users are much more likely to be
relying on the binary than installing libgit2 with Homebrew and building
git2r from source).

https://docs.travis-ci.com/user/installing-dependencies/#installing-packages-on-macos