Wrap version control check in p tags. [skip ci] [ci skip]
The p tags are automatically and properly added to the other checks when
inserted inside the body-panel. Presumably because of the pre tags to
display the git status in a code chunk, this messes up the p tags.
Adding them manually for all chunks (in format_check()) only created
duplicate empty p tags. Thus the solution was to add p tags only to the
version control check.