Remove unnecessary warnings.
* wflow_start() sent a warning about an existing Git repo even when the
user explicitly set existing=TRUE. This is an artifact. Now that the
issue of an existing Git repo is caught very early in the function, this
is no longer necessary.

* wflow_paths() sent a warning if the website directory didn't exist.
This doesn't actually hurt anything. Running wflow_build() or
wflow_publish() will create the directory.

* wflow_publish() sent a warning if the website directory didn't exist
because it tried unsuccessfully to create a backup copy. Now
wflow_build() does not attempt to copy if it doesn't exist (also made
the improvement that it doesn't copy if it's a dry run).