More winbuilder exerpimenting. [skip ci] [ci skip]
What I learned:

* fs::dir_exists() returns FALSE if the directory has a trailing slash.
This only happens on winbuilder. It doesn't occur on my local Windows 10
machine or on AppVeyor. This is what was causing the errors with
wflow_rename()/wflow_use_gitlab(). These involve
renaming a directory and the early check with fs::dir_exists() would
fail.

* The tests are run in
d:/RCompile/CRANguest/R-release/workflowr.Rcheck/tests_x64/testthat
and the temporary directory is D:/temp/RtmpoNs5r5/file7cb017ec743. They
are both on the D drive, but notice the difference in the case of the D
drive. Both getwd() and fs::path_wd() return the current working
directory with a lower case d.

* workflowr:::relative() appears to return the correct relative path
from the test directory to a temporary file. It goes up 6 levels to D:/
and then goes forward starting with temp/, e.g.
../../../../../../temp/RtmpoNs5r5/file7cb027113cf9. I think this is the
correct relative path because fs::file_exists returns TRUE.

* fs::path_rel() does not appear to return the correct relative path to
a temporary file. It goes up 7 directories and then starts with D:/.
Basically it behaves as if there is a root above all the drives,
e.g.  ../../../../../../../D:/temp/RtmpoNs5r5/file7cb017ec743. I think
this is incorrect because fs::file_exists() returns FALSE. The problem
appears to arise from the fast that fs::path_wd() returns d:/ and
fs::file_temp() returns D:/, and then fs::path_rel() doesn't recognize
that these are identical on a case-insensitive file system like Windows.

* Because wflow_git_commit() starts by converting the files to relative
paths, this is likely causing the problems on winbuilder. I like
returning relative paths to the user, so I need to try to keep the paths
absolute as much as possible expect for message and return() statements
to see if this improves the situation. The problem will be
wflow_publish() which calls wflow_git_commit() and wflow_build().

* R.home() is D:/RCompile/recent/R-35~1.2

* fs::path_home() is C:/Users/CRAN

* fs::path_home_r() is C:/Users/CRAN/Documents

* wflow_update() fails because _workflowr.yml is not committed. It is
still an untracked file. Maybe this is because it is the only file that
is completely new to Git? All the others are being modified or deleted.

* wflow_use_gitlab() fails because it commits .gitlab-ci.yml (a new
file) but not analysis/_site.yml (an existing file). The variable
output_dir had been changed to ../public, so it should have had changes
to be committed with Git.

* wflow_use_github() fails when wflow_publish() is run afterwards.
wflow_git_commit_ claims there are no HTML files to commit. Running
wflow_build() beforehand confirmed that the Rmd files are built. Maybe
it is looking for the HTML files in the wrong directory? Or maybe
relative paths are causing problems?