Create git2r_add to handle all edge cases.
git2r::add() is a constant source of frustration because the file paths
it accepts depend on various combinations of factors that behave in
unpredictable ways.

Paths can be passed in multiple ways: 1) relative to the current working
directory, 2) relative to the Git directory (which isn't always the same
as the former), and 3) absolute paths.

The following affect which type of path can be used:

* Operating system, especially Windows.

* Whether the file is being modified, deleted, or added for the first
time. wflow_update() was failing because it added the modified files
but couldn't add the untracked file _workflowr.yml. A deleted file can't
be added with 1). It has to be relative or absolute. The motivation for
the commit below was to allow wflow_remove() to reliably remove files
from the Git repo.

https://github.com/jdblischak/workflowr/commit/c28bf02067d0941c877ae9866849317fbf5ae3da

* The version of git2r or libgit2. This has changed over time. The
commit above solved the problem for the previous release, and it stil
works for the most part. It mainly just choked on winbuilder.

* Arguments like "." have a special meaning (git add .), so shouldn't be
altered.