Avoid directories that may actually exist in unit tests.
winbuilder capitalized Temp. I know that D:/Temp exists, but it also
capitalized it when it was passed C:/temp. Unclear if this is a bug or
expected because I don't know if C:/Temp exists. To be safe, just avoid
this altogether. This won't actually affect end users on Windows since
the file system is case-insensitive.