Fix print.wflow_rename() error on winbuilder with r-devel.
Everything else ignored this, but sometimes I passed mutiple arguments
to is.na() in combination with &&, which isn't a great idea.

The reason is that I'm trying to catch when there is no Git repo
present. In that case, there will be files_git and git=TRUE since that
is default. But in every other case, files_git will be at least length
two since a file was renamed, which is what caused the problem. Wrapping
is.na() in all() solves both cases.