Setup a pandoc-less testing environment on CircleCI.
I use skip_on_cran() to avoid running pandoc on the CRAN servers.
However, all of my automated tests have pandoc installed, and so does
rhub and winbuilder. This CI will ensure that I don't forget to skip any
tests.

workflowr 1.2.0 passed the initial CRAN tests on Windows and Debian.
Hoewver, it failed on r-patched-solaris-x86 because of a call to
wflow_publish().

https://dev.appsilon.com/continuous-integration-for-your-private-r-projects-with-circleci/
https://hub.docker.com/r/rocker/r-apt
https://circleci.com/docs/2.0/configuration-reference/
https://circleci.com/docs/2.0/configuration-reference/#full-example