Only test knitr option indent for markdown output.
The HTML result varies based on the version of pandoc. This option is
only supposed to be used for markdown output, so testing the
intermediate markdown file is sufficient to demonstrate that
wflow_html() is not interfering with the standard behavior that a user
would expect from knitr/rmarkdown.

Related to #149.