Make add_bibliography() tests more flexible.
For some reason, rmarkdown 1.7 with pandoc 2.7 added a role attribute to
the refs div tag. This broke the regex that searched the HTML for the
position of this bibliography.

<div id="refs" role="doc-bibliography">

I didn't need to change the code. Pandoc still expects the user to
manually include <div id="refs"></div> to specify a location for the
bibliography. At least as far as I can tell. I tried manually assiging
the role attribute, and this resulted in pandoc then adding a duplicate
refs div but with a class attribute.

<div id="refs" class="references">

If I updated add_bibliography() to use a more flexible regex and
manually added the role attribute, then pandoc added a duplicate role
attribute. Thus I think it is clear that this shouldn't be added by the
user. And thus doesn't need to be detected by workflowr. This is an edge
case anyways, for users that manually added a bibliography before this
was supported by default.