Handle spaces in chunk names for HTML targets.
If a chunk name has a space in it, then the filename has a space in it.
Unfortunately this also meant that the data-target and div id in the
HTML included the space. Since spaces aren't allowed, this broke the
collapsible button that displays the table of past versions.

This update removes spaces from the chunk names and also makes sure to
avoid clashes between chunks in the same document that would be the same
after converting spaces to dashes.

Discovered here:

https://brimittleman.github.io/threeprimeseq/total.nuc.rnaseq.html#mapped_reads