Detect absolute paths in strings.
Will be used in a reproducibility check for wflow_html().

The difficulty is in allowing spaces in file paths, which complicates
the regular expressions. I had to anchor them by requiring they be
surrounded by quotations marks, e.g. a string in the R code, or by
parentheses, e.g. a link in Markdown syntax.

Even this didn't allow finding multiple absolute paths on the same line.
Getting this to work required changing .+ to the non-greedy .+? to stop
at the first appearance of the next quotation mark or parentheses.

https://stackoverflow.com/a/171483/2483477