Find project directory to determine internal paths.
This is complicated by the fact that wflow_html() does not require the
Rmd file to be in a workflowr project. Thus I created an ad-hoc search
for the likely root of the project. For now I am avoiding adding the
dependency here, but many workflowr uesrs already use it, so it may be
worth just requiring it.