Don't attemt to report absolute paths in Markdown links.
This gets too complicated because Markdown links always need to be
relative to the docs/ directory. Since a wflow_html() Rmd file may not
even be part of a workflowr project, this may not even make sense. And
even for workflowr projects, trying to explain the difference between
relative paths to docs/ and to the knit directory would be very
confusing for the exact users this check is trying to help. An absolute
link in an HTML link is inconvenient, but doesn't affect the code
execution.