Display original Windows absolute paths.
Backslashes are a huge pain. "\\" in R is output as "\". Thus to display
the double backslash in the report, my code has to use "\\\\".
Furthermore, since that is an escape sequence, it needs to be doubled
when used in a regex like the stringr functions. Thus this code looks
disgusting!

Also, I occasionally got some pandoc errors when building the file
interactively in RStudio. It is similar to the errors in the issue
below. If this turns out to be too problematic, I'll just remove the
support for displaying backslashes in the report.

https://github.com/rstudio/rstudio/issues/3661