Fix bug due to unstaged conflicting change when pulling.
Identified by Peter. Both workflowr and git2r can handle conflicting
changes when the changes are committed. However, unstaged changes in
conflicting files causes the git_merge_result object returned by git2r
to behave unexpectedly. I opened an Issue on git2r to try to handle this
upstream:

https://github.com/ropensci/git2r/issues/389

For now I catch this in workflowr and try to provide messaging similar
to Git. I didn't display the problematic files because technically the
print method does not have access to the location of the Git repo.