Pin rocker/r-ver:devel to the version from 2 weeks ago.
The devel image was recently updated to use Debian 10 (buster)
and then libicu63.

https://github.com/rocker-org/rocker-versioned/commit/b1d2b1b1f8b338b66fb8f855899413a36ac60462#diff-8e87d872cd4b82b711d6ad5bc2fc01c9
https://github.com/rocker-org/rocker-versioned/commit/b4a4b0eec67785ca8d9588b3f1e227cdf9f59208#diff-8e87d872cd4b82b711d6ad5bc2fc01c9R29

This causes littler to crash unexpectedly (it doesn't mention that
it depends on icu):

r: error while loading shared libraries: libicuuc.so.57: cannot open
shared object file: No such file or directory
Exited with code 127

I am pretty sure this is due to the cache being outdated. stringi
is built with libicu57 and not libicu63, which is what is installed.
Unfortunately there is no way to automatically invalidate the cahce
due to an OS update:

https://support.circleci.com/hc/en-us/articles/360004937834-Resolving-Ruby-Jessie-to-Stretch-issue-cannot-open-shared-object-file-cache-problem

I'm still a little confused why Rscript worked but not littler,
since the latter does not appear to rely on icu. Unclear why it
would be trying to load stringi.