Speed up wflow_publish() by removing unneeded call to wflow_status().
This reduced republishing a file with no code from ~2 sec to ~1.5 sec.
Note that this repo had 18 Rmd files.

This also made the detection of conflicts more robust. Before it
only detected conflicts in Rmd files. But this is insufficient. The
commit will fail even if non-Rmd files (even files not passed to
wflow_git_commit()) will cause the commit to fail.

Note that it is important for workflowr to catch this error b/c
the message returned by git2r::commit() upon failure is cryptic:

> Error in 'git2r_commit': cannot create a tree from a not fully merged
index.