Warn user if knit_root_dir is absolute path.
Note that this check is only performed for knit_root_dir defined
in _workflowr.yml. With the current implementation, it would be
difficult to test for an absolute path only in the knit_root_dir
defined in the YAML header. As this is an advanced technique and
generally discouraged anyways, I don't think it is worth the
effort to test for it.