Require stringr >= 1.3.0 to prevent wflow_start() infinite recursion error.
I remember seeing this error at the BSD workshop back in April 2019,
but had no idea what was causing it. I happened to reproduce it
when in the RStudio Server hosted in the Docker image rocker/verse
(I wanted to test the httpuv error with the Viewer pane). I got
the same error message as some of the students:

wflow_start("test")
Error: evaluation nested too deeply: infinite recursion /
options(expressions=)?
Error during wrapup: evaluation nested too deeply: infinite recursion /
options(expressions=)?

It happened no matter the directory, or using the RStudio project
template.

The error was caused by the recursive function
obtain_existing_path(). It didn't have a stop criteria if it was
passed nonsense like NA, so it went on recursively until it errored.

The NA came from absolute() -> toupper_win_drive(). stringr 1.2.0
introduced the ability to use a function to perform the replacement.
This is how toupper_win_drive() works, it uses toupper. The problem
is that if there is no match, instead of returning the original
string, it returns NA.

This was fixed in stringr 1.3.0 in PR #164. The NEWS bullet is
quite subtle for how much it changes the behavior.

https://github.com/tidyverse/stringr/pull/164