Remove methods as a dependency.
This was originally added for methods::as() when dealing with
the git2r S4 classes. Now that those have been completely dropped
by requiring a minimum git2r that uses S3, methods is no longer
needed.

https://github.com/jdblischak/workflowr/commit/66423a93e1886e6599ceec7f8dd0ebea06434a12

Suprisingly, this NOTE only occurred when testing with
R 3.2.5. More recent versions of R had no issue with it.

* checking dependencies in R code ... NOTE
Namespace in Imports field not imported from: ‘methods’