Set more minimum version requirements for dependencies.
I identified these while testing workflowr with R 3.2.5 in
the Docker image rocker/r-ver:3.2.5. Since it uses an
MRAN snapshot from 2016, it identified issues with
functions not yet available in those past versions.

https://mran.microsoft.com/snapshot/2016-05-03

clipr::clipr_available() was added in 0.7.0:
https://github.com/mdlincoln/clipr/blob/master/NEWS.md#clipr-070

rstudioapi::sendToConsole and getSourceEditorContext
were added in 0.6:
https://github.com/rstudio/rstudioapi/blob/235985d48266152de9efa8e871caa331f760f777/NAMESPACE

testthat::expect_setequal() was added in 2.0.0:
https://github.com/r-lib/testthat/blob/master/NEWS.md#testthat-200

shiny::textAreaInput was added in 0.14:
https://github.com/rstudio/shiny/blob/master/NEWS.md#shiny-014

withr::local_ was added in 2.0.0:
https://github.com/r-lib/withr/blob/master/NEWS.md#withr-200