Set httpuv as a dependency for use with wflow_use_github().
wflow_use_github() uses the httr package for oauth authentication
with GitHub. This is an important step specifically for novices
that may not be familiar/comfortable with GitHub. Unfortunately
httpuv is not automatically installed with workflowr because
httr lists httpuv as a suggested package, and has code to handle
this:

https://github.com/r-lib/httr/search?q=httpuv&unscoped_q=httpuv

If this were a developer-only focused package, I would do the
same. But the users that will rely on this the most need it
to simply be installed. Thus I listed it as a required dependency.

Supposedly RStudio Desktop requires httpuv 1.2.0 in order to
use the RStudio Viewer pane. However, I think this must be
context dependent. I uninstalled httpuv completely from my
Ubuntu machine running RStudio 1.2.1335, and I had no
problem using the RStudio Viewer pane.

https://github.com/rstudio/rstudio/blob/08eb8fd120c3922e405d26074ee1f0c3a02cc5d1/src/cpp/session/modules/viewer/SessionViewer.cpp#L335

However, I couldn't get httpuv 1.2.0 to compile. Thus I
chose the oldest version that I could still compile,
which was httpuv 1.2.2. This makes sense since this is when
support for Rcpp 0.11.0 was added:

https://cran.r-project.org/web/packages/httpuv/news/news.html

R CMD check doesn't complain about this dependency that it
doesn't look like I use. Presumably R 3.2.5 will complain since
it did about the methods package.

https://github.com/jdblischak/workflowr/commit/e159c4bf0219d811080dd169272fa6a8b07db8e6

But maybe they got rid of that NOTE because of use cases
like this. Who knows?