Prevent spurious errors due to Windows rounding of dates.
Sometimes these tests would fail on Windows due to very small
differences, which is ridiculous since I am testing the date and
not the time. But to prevent it, I test it with a tolerance using
all.equal().

https://ci.appveyor.com/project/jdblischak/workflowr/builds/28394053#L1129

Running the tests in 'tests/testthat.R' failed.
Last 13 lines of output:
  versions_f2$Date not identical to
c(as.Date(as.POSIXct(c4$author$when)),
as.Date(as.POSIXct(c2$author$when))).
  Mean relative difference: 5.496317e-05

  -- 3. Failure: get_versions_df returns data frame of commits for
file(s) (@test-
  versions_f1_f2$Date not identical to c(...).
  Mean relative difference: 5.496317e-05

  == testthat results
===========================================================
  [ OK: 1602 | SKIPPED: 8 | WARNINGS: 0 | FAILED: 3 ]
  1. Failure: get_versions_df returns data frame of commits for file(s)
(@test-report.R#45)
  2. Failure: get_versions_df returns data frame of commits for file(s)
(@test-report.R#54)
  3. Failure: get_versions_df returns data frame of commits for file(s)
(@test-report.R#63)