Avoid NOTE from R CMD check.
wflow_use_github() requires httpuv::startServer() to be called
internally from httr to enable oauth authentication. Thus I
ensure it is installed by including it as an imported
dependency.

https://github.com/r-lib/httr/blob/6b84c576f866956e01835e1b6f02c55c8012d4e0/R/oauth-listener.r#L46

Howver, this sometimes generates the following
NOTE from R CMD check (e.g. on rhub but not win-builder):

* checking dependencies in R code ... NOTE
Namespace in Imports field not imported from: ‘httpuv’
  All declared Imports should be used.

which can be avoided by importing a function from that package
in the NAMESPACE.

https://community.rstudio.com/t/how-should-a-meta-package-handle-this-note-all-declared-imports-should-be-used/23400/3
https://github.com/tidyverse/tidyverse/blob/a720dcd73d9e3fc0ec86317bc0abaf8f0077e8bd/R/tidyverse.R#L4