Fix conversion of git_time to character string of date.
By first converting to a character, the timezone information was
lost. Since git_time records the time as GMT, the final time was
off when it was converted to EDT. Thus the date was wrong, but the
tests usually passed since I made the same mistake in both
get_versions_df() and in the test. For some reason only the r-hub
macOS machine consistently failed enough for me to discover and
debug the issue.

Dealing with Dates in R can be a real pain. I got the do.call()
line from this SO answer, which was the only way I could get a
vector of Dates.

https://stackoverflow.com/a/13859997