Parsing is too unstable so hard-code minimum required versions.
During R CMD check, the DESCRIPTION file must be edited. It was
failing because it couldn't find the package "callr," (with the
trailing comma). I was able to reproduce a similar error by putting
callr and fs on the same line in the DESCRIPTION file.

Thus it is unfortunate that this duplication will require updating
in 2 places in the future, but importantly it is robust to any
modifications to DESCRIPTION.

Issue #178