Pass value of sysgit to last_commit_time_sysgit().
Need to respect value of workflowr.sysgit. A user may choose to use
a Git executable different from the one that is first on the PATH.

Had to include the unused ... argument so that sysgit could be passed
to last_commit_time_git2r() without triggering an error. Also added
... to last_commit_time_sysgit() to make the functions signature more
similar (and also for any potential future use).