Update description of past versions of the repository.
Also, I removed the direct link to the Rmd in the repository check. That
was too tricky because there are so many scenarios to handle (e.g. if
the file has staged/unstaged changes, the link would be meaningless).
Because workflowr::wflow_html() is designed to be flexible enough to be
used outside of a workflowr project, the downside is that it's more
difficult to ascertain the status of the Rmd file (e.g. I can't just run
wflow_status("file.Rmd") because that assumes a workflowr project).