Fix Windows-specific bug in the versions table.
I was using a path relative to the root of the Git repo. This should
work because in general git2r supports these types of input paths.
However, in this case, the relative path on Windows gets converted to
a non-existent absolute by path via commits() -> sanitize_path() ->
normalizePath(). Thus my solution is to pass absolute paths to
git2r::commits(), and then switch to relative paths for displaying
to users in the version table.

https://github.com/ropensci/git2r/issues/394#issuecomment-585989444