Use xfun::isFALSE() instead of isFALSE().
isFALSE() was introduced in R 3.5.0. Instead of using the backports
package, which would add a dependency, I chose to use the xfun package,
which also has an implementaiton of isFALSE(), and is a required
dependency of knitr, so it isn't adding a new dependency for workflowr.

https://cran.r-project.org/doc/manuals/r-release/NEWS.html
https://github.com/yihui/knitr/commit/48196771e899e0e94a73e0faf02b47b771d153fa

Note that the update to isTRUE() in R 3.5.0 shouldn't affect my code
since I am checking TRUE, FALSE, NA, and NULL.