Avoid false positive test failures for pandoc title warning.
The tests that check for pandoc warnings required the calls to be
absolutely silent. This is misleading since any random message sent to
stderr would cause the tests to fail, even though nothing had changed as
far as this behavior.

This was happening on the weekly build "future", which used a slightly
different strategy for installing system dependencies compared to the
"devel" build. As far as I can tell that is what caused the difference.

https://circleci.com/gh/workflowr/workflowr-pkg-tests/686

The warning came from grSoftVersion():

Warning message:
In grSoftVersion() :
  unable to load shared object '/usr/local/lib/R/modules//R_X11.so':
  libSM.so.6: cannot open shared object file: No such file or directory

I fixed it by installing libsm6 and then libxt6.

Once I confirm this test update fixes things, then I'll update the
"future" build so that it is more similar to the other builds.