Fail early if wflow_publish() was not told which files to publish.
This was discovered during the UChicago QBio workshop. Users expected
wflow_publish() to behave similar to wflow_build(), in that it would
automatically do what it needed without explicit direction. Since
publishing a file is much hard to undo than building it, users need to
explicitly specify the files to publish. Since there is no use case for
running wflow_publish() without instructing it to publish some files, I
updated it to fail early in this situation. The only way to obtain the
old behavior is to set dry_run=TRUE.