Require minimum of pandoc 1.14. [skip ci]
The rmarkdown package switched to requiring pandoc >= 1.14 a few months ago.
https://github.com/rstudio/rmarkdown/blame/master/DESCRIPTION

This was to support table of contents in Word documents. This has nothing to do
with workflowr.
https://github.com/rstudio/rmarkdown/commit/b8dcd87bf2a5ea7f070473d66e433b54e33ce765

However, pandoc 1.14 was released on 2015-05-27, and the previous minimum (and
current for workflowr) of 1.12.3 was released on 2014-01-10.
https://pandoc.org/releases.html

The oldest version I actually regularly test is 1.19, so I think it would be
good to bump up the minimum.
https://github.com/workflowr/workflowr-pkg-tests/blob/11b2189906d5513008bdd54172fdb066c2b72b0b/.circleci/config.yml#L268