Don't build vignettes for pandocless job
Now that knitr moved the markdown package to Suggests, the
rmarkdown engine can't fall back v1 to create the vignettes.
The purpose of this job is to catch any tests that don't first
check for the presence of pandoc, not to check the vignettes,
so it's fine to simply skip this.

https://github.com/yihui/knitr/issues/1864